+++ sdk/core/azure-core-amqp/src/amqp/connection.cpp
@@ -677,7 +677,9 @@ namespace Azure { namespace Core { namespace Amqp { namespace _detail {
             context);
         if (std::get<0>(result) != CbsOperationResult::Ok)
         {
-          throw std::runtime_error("Could not put Claims Based Security token.");
+          throw Azure::Core::Credentials::AuthenticationException(
+              "Could not authenticate client. Error Status: "
+              + std::to_string(std::get<1>(result)) + " reason: " + std::get<2>(result));
         }
         Log::Stream(Logger::Level::Verbose) << "Close CBS object";
         claimsBasedSecurity->Close(context);