+++ sdk/core/azure-core-amqp/test/ut/claim_based_security_tests.cpp
@@ -75,6 +75,7 @@ namespace Azure { namespace Core { namespace Amqp { namespace Tests {
       EXPECT_EQ(CbsOpenResult::Error, cbs.Open());
     }
   }
+
   TEST_F(TestCbs, CbsOpen)
   {
     MessageTests::AmqpServerMock mockServer;
@@ -103,6 +104,33 @@ namespace Azure { namespace Core { namespace Amqp { namespace Tests {
     }
     mockServer.StopListening();
   }
+
+  TEST_F(TestCbs, CbsCancelledOpen)
+  {
+    MessageTests::AmqpServerMock mockServer;
+
+    mockServer.EnableTrace(false);
+
+    ConnectionOptions options;
+    options.Port = mockServer.GetPort();
+    options.EnableTrace = true;
+    options.ContainerId = testing::UnitTest::GetInstance()->current_test_info()->test_case_name();
+    Connection connection("localhost", nullptr, options);
+    Session session{connection.CreateSession()};
+
+    mockServer.StartListening();
+
+    {
+      GTEST_LOG_(INFO) << "Create CBS object.";
+      ClaimsBasedSecurity cbs(session);
+      Azure::Core::Context openContext;
+      openContext.Cancel();
+      CbsOpenResult openResult = cbs.Open(openContext);
+      EXPECT_EQ(CbsOpenResult::Cancelled, openResult);
+    }
+    mockServer.StopListening();
+  }
+
 #endif // !defined(AZ_PLATFORM_MAC)
 
 #if !defined(AZ_PLATFORM_MAC)
@@ -157,9 +185,46 @@ namespace Azure { namespace Core { namespace Amqp { namespace Tests {
         GTEST_LOG_(INFO) << "Open Completed.";
 
         mockServer.ForceCbsError(true);
-        EXPECT_ANY_THROW(
-            auto putResult = cbs.PutToken(
-                Azure::Core::Amqp::_detail::CbsTokenType::Sas, "of one", "stringizedToken"););
+        auto putResult = cbs.PutToken(
+            Azure::Core::Amqp::_detail::CbsTokenType::Sas, "of one", "stringizedToken");
+        EXPECT_EQ(CbsOperationResult::Failed, std::get<0>(putResult));
+        cbs.Close();
+      }
+
+      mockServer.StopListening();
+    }
+  }
+
+  TEST_F(TestCbs, CbsOpenAndPutCancelled)
+  {
+    {
+      MessageTests::AmqpServerMock mockServer;
+
+      ConnectionOptions options;
+      options.Port = mockServer.GetPort();
+      options.EnableTrace = true;
+      Connection connection("localhost", nullptr, options);
+      Session session{connection.CreateSession()};
+
+      mockServer.StartListening();
+
+      {
+        ClaimsBasedSecurity cbs(session);
+
+        EXPECT_EQ(CbsOpenResult::Ok, cbs.Open());
+        GTEST_LOG_(INFO) << "Open Completed.";
+
+        Azure::Core::Context putContext;
+        putContext.Cancel();
+
+        //        mockServer.ForceCbsError(true);
+        EXPECT_EQ(
+            CbsOperationResult::Cancelled,
+            std::get<0>(cbs.PutToken(
+                Azure::Core::Amqp::_detail::CbsTokenType::Sas,
+                "of one",
+                "stringizedToken",
+                putContext)));
 
         cbs.Close();
       }
@@ -167,5 +232,6 @@ namespace Azure { namespace Core { namespace Amqp { namespace Tests {
       mockServer.StopListening();
     }
   }
+
 #endif // !defined(AZ_PLATFORM_MAC)
 }}}} // namespace Azure::Core::Amqp::Tests