+++ sdk/core/azure-core-amqp/CHANGELOG.md
@@ -6,12 +6,14 @@
 
 ### Breaking Changes
 
-- Claims Based Security authentication now throws `Azure::Core::Credentials::AuthenticationException` instead of `std::runtime_error`.
+- Claims Based Security authentication now longer throws a `std::runtime_error`, and instead follows the pattern of the rest of the AMQP library and returns an error.
+- Authentication now throws `Azure::Core::Credentials::AuthenticationException` instead of `std::runtime_error`.
+- Added `Cancelled` status to `CbsOperationResult` and `ManagementOperationStatus`.
 
 ### Bugs Fixed
 
-- [#5284](https://github.com/Azure/azure-sdk-for-cpp/issues/5284): [azure-identity][azure-messaging-eventhubs] Impossible to catch exception resulting in SIGABRT signal.
-- [#5297](https://github.com/Azure/azure-sdk-for-cpp/issues/5297): Enabled multiple simultaneous `ExecuteOperation` calls.
+- [[#5284]](https://github.com/Azure/azure-sdk-for-cpp/issues/5284) [azure-identity][azure-messaging-eventhubs] Impossible to catch exception resulting in SIGABRT signal.
+- [[#5297]](https://github.com/Azure/azure-sdk-for-cpp/issues/5297): Enabled multiple simultaneous `ExecuteOperation` calls.
 - Fixed crash when Link Detach message is received while link is being destroyed.
 
 ### Other Changes